requirements.txt
python-binance
python-dotenv
pandas
numpy
matplotlib
